revert: Revert file refactor
Reverted file refactor
