+++ src/components/PublicationRequestSections/PublicationType/BookDetails.js
