+++ src/components/PublicationRequestSections/PublicationType/index.js
@@ -1,0 +1,2 @@
+export { default as BookDetails } from './BookDetails';
+export { default as JournalDetails } from './JournalDetails';
