+++ src/components/PublicationRequestSections/index.js
@@ -4,8 +4,8 @@ export { default as Publication } from './Publication';
 export { default as PublicationStatus } from './PublicationStatus';
 export { default as RequestContact } from './RequestContact';
 export { default as RequestInfo } from './RequestInfo';
-export { default as BookDetails } from './BookDetails';
-export { default as JournalDetails } from './JournalDetails';
+export { default as BookDetails } from './PublicationType';
+export { default as JournalDetails } from './PublicationType';
 export { default as PartyInfo } from './PartyInfo';
 export { default as Correspondence } from './Correspondence';
 export { default as Agreement } from './Agreement';
